requirements.txt
google-adk
google-cloud-firestore
requests
fastapi
uvicorn
python-dotenv
google-generativeai
